Fix all syncSwap calculations
