--- src/features/on-chain/calculation-manager/providers/dexes/zksync/sync-swap/sync-swap-factory-abi.ts
@@ -1,28 +1,0 @@
-import { AbiItem } from 'web3-utils';
-
-export const syncSwapFactoryAbi: AbiItem[] = [
-    {
-        inputs: [
-            {
-                internalType: 'address',
-                name: '',
-                type: 'address'
-            },
-            {
-                internalType: 'address',
-                name: '',
-                type: 'address'
-            }
-        ],
-        name: 'getPool',
-        outputs: [
-            {
-                internalType: 'address',
-                name: '',
-                type: 'address'
-            }
-        ],
-        stateMutability: 'view',
-        type: 'function'
-    }
-];
