+++ src/features/on-chain/calculation-manager/providers/dexes/zksync/sync-swap/sync-swap-provider.ts
@@ -1,6 +1,6 @@
 import BigNumber from 'bignumber.js';
 import { NotSupportedTokensError } from 'src/common/errors';
-import { PriceToken, PriceTokenAmount, wrappedNativeTokensList } from 'src/common/tokens';
+import { PriceToken, PriceTokenAmount, Token, wrappedNativeTokensList } from 'src/common/tokens';
 import { combineOptions } from 'src/common/utils/options';
 import { BLOCKCHAIN_NAME, EvmBlockchainName } from 'src/core/blockchain/models/blockchain-name';
 import { createTokenNativeAddressProxy } from 'src/features/common/utils/token-native-address-proxy';
@@ -13,9 +13,9 @@ import {
 import { evmProviderDefaultOptions } from 'src/features/on-chain/calculation-manager/providers/dexes/common/on-chain-provider/evm-on-chain-provider/constants/evm-provider-default-options';
 import { EvmOnChainProvider } from 'src/features/on-chain/calculation-manager/providers/dexes/common/on-chain-provider/evm-on-chain-provider/evm-on-chain-provider';
 import { SyncSwapTrade } from 'src/features/on-chain/calculation-manager/providers/dexes/zksync/sync-swap/sync-swap-trade';
-import { PathFactoryV2 } from 'src/features/on-chain/calculation-manager/providers/dexes/zksync/sync-swap/utils/path-factory-v2';
-import { PoolFactory } from 'src/features/on-chain/calculation-manager/providers/dexes/zksync/sync-swap/utils/pool-factory';
-import { findBestAmountsForPathsExactIn } from 'src/features/on-chain/calculation-manager/providers/dexes/zksync/sync-swap/utils/router';
+import { SyncSwapFactory } from 'src/features/on-chain/calculation-manager/providers/dexes/zksync/sync-swap/utils/sync-swap-factory';
+import { SyncSwapPathFactory } from 'src/features/on-chain/calculation-manager/providers/dexes/zksync/sync-swap/utils/sync-swap-path-factory';
+import { SyncSwapRouter } from 'src/features/on-chain/calculation-manager/providers/dexes/zksync/sync-swap/utils/sync-swap-router';
 import { RoutePools } from 'src/features/on-chain/calculation-manager/providers/dexes/zksync/sync-swap/utils/typings';
 
 export class SyncSwapProvider extends EvmOnChainProvider {
@@ -37,14 +37,17 @@ export class SyncSwapProvider extends EvmOnChainProvider {
             throw new NotSupportedTokensError();
         }
 
-        const paths = PathFactoryV2.findAllPossiblePaths(
+        const paths = SyncSwapPathFactory.findAllPossiblePaths(
             from.address,
             toToken.address,
             availablePools
         );
-        const filteredPaths = PathFactoryV2.getBestPath(paths, from.stringWeiAmount);
+        const filteredPaths = await SyncSwapPathFactory.getBestPath(paths, from.stringWeiAmount);
 
-        const bestRoute = await findBestAmountsForPathsExactIn(filteredPaths, from.stringWeiAmount);
+        const bestRoute = await SyncSwapRouter.findBestAmountsForPathsExactIn(
+            filteredPaths,
+            from.stringWeiAmount
+        );
 
         const fromAddress =
             options?.useProxy || this.defaultOptions.useProxy
@@ -58,36 +61,22 @@ export class SyncSwapProvider extends EvmOnChainProvider {
             from,
             wrappedNativeTokensList[from.blockchain]!.address
         );
-        // const toProxy = createTokenNativeAddressProxy(
-        //     toToken,
-        //     wrappedNativeTokensList[from.blockchain]!.address
-        // );
-        //
+
         const { fromWithoutFee, proxyFeeInfo } = await this.handleProxyContract(
             fromProxy,
             fullOptions
         );
-        //
-        // const routePools = await this.fetchRoutePools(fromWithoutFee.address, toProxy.address);
-        // const poolData = this.getPoolData(routePools);
-        // if (
-        //     poolData.reserveA === '0' ||
-        //     poolData.reserveB === '0' ||
-        //     compareAddresses(poolData.pool, EvmWeb3Pure.EMPTY_ADDRESS)
-        // ) {
-        //     throw new NotSupportedTokensError();
-        // }
-        //
-        // const weiAmountOut = await this.getAmountOut(
-        //     fromWithoutFee.weiAmount,
-        //     fromWithoutFee.address,
-        //     poolData
-        // );
 
         const to = new PriceTokenAmount({
             ...toToken.asStruct,
             weiAmount: new BigNumber(bestRoute.amountOut.toString())
         });
+
+        const transitAddresses = bestRoute.pathsWithAmounts[0]!.stepsWithAmount.slice(1).map(
+            step => step.tokenIn
+        );
+
+        const transitTokens = await Token.createTokens(transitAddresses, from.blockchain);
 
         const tradeStruct = {
             from,
@@ -99,133 +88,30 @@ export class SyncSwapProvider extends EvmOnChainProvider {
             fromWithoutFee,
             withDeflation: fullOptions.withDeflation,
             usedForCrossChain: fullOptions.usedForCrossChain,
-            path: [from, toToken],
+            path: [from, ...transitTokens, toToken],
             bestPathWithAmounts: bestRoute
         };
 
         return new SyncSwapTrade(tradeStruct, fullOptions.providerAddress);
     }
 
-    // private fetchRoutePools(tokenFrom: string, tokenTo: string): Promise<unknown> {
-    //     const web3Public = Injector.web3PublicService.getWeb3Public(BLOCKCHAIN_NAME.ZK_SYNC);
-    //     return web3Public.callContractMethod(
-    //         '0x5c07e74cb541c3d1875aeee441d691ded6eba204',
-    //         routerSupportAbi,
-    //         'getRoutePools',
-    //         [
-    //             tokenFrom,
-    //             tokenTo,
-    //             [
-    //                 '0xf2dad89f2788a8cd54625c60b55cd3d2d0aca7cb',
-    //                 '0x5b9f21d407f35b10cbfddca17d5d84b129356ea3'
-    //             ],
-    //             [tokenFrom, tokenTo],
-    //             '0xbb05918e9b4ba9fe2c8384d223f0844867909ffb',
-    //             this.walletAddress || EvmWeb3Pure.EMPTY_ADDRESS
-    //         ]
-    //     );
-    // }
-    //
-    // public getPoolData(poolsInfo: unknown): unknown {
-    //     const { poolsBase } = poolsInfo as { poolsBase: any[] };
-    //     const isFirstPoolReserveABetter = new BigNumber(poolsBase[0]?.reserveA || 0).gt(
-    //         new BigNumber(poolsBase[1]?.reserveA || 0)
-    //     );
-    //     const isFirstPoolReserveBBetter = new BigNumber(poolsBase[0]?.reserveB || 0).gt(
-    //         new BigNumber(poolsBase[1]?.reserveB || 0)
-    //     );
-    //
-    //     return isFirstPoolReserveABetter && isFirstPoolReserveBBetter
-    //         ? poolsBase[0]!
-    //         : poolsBase[1]!;
-    // }
-    //
-    // private async getAmountOut(
-    //     amountIn: BigNumber,
-    //     fromAddress: string,
-    //     poolInfo: PoolInfo
-    // ): Promise<BigNumber> {
-    //     const [reserveFrom, reserveTo, fee, tokenFromAddress] = compareAddresses(
-    //         poolInfo.tokenA,
-    //         fromAddress
-    //     )
-    //         ? [poolInfo.reserveA, poolInfo.reserveB, poolInfo.swapFeeAB, poolInfo.tokenA]
-    //         : [poolInfo.reserveB, poolInfo.reserveA, poolInfo.swapFeeBA, poolInfo.tokenB];
-    //     const maxFee = new BigNumber(100_000);
-    //
-    //     // Stable swap
-    //     if (poolInfo.poolType === '2') {
-    //         const { fromPrecisionMultiplier, toPrecisionMultiplier } = await this.getPoolPrecision(
-    //             poolInfo.pool,
-    //             tokenFromAddress
-    //         );
-    //         const out = getAmountOutStable({
-    //             amountIn: BigNumberEthers.from(amountIn.toFixed()),
-    //             reserveIn: BigNumberEthers.from(reserveFrom),
-    //             reserveOut: BigNumberEthers.from(reserveTo),
-    //             swapFee: BigNumberEthers.from(fee),
-    //             A: BigNumberEthers.from(1000),
-    //             tokenInPrecisionMultiplier: BigNumberEthers.from(fromPrecisionMultiplier),
-    //             tokenOutPrecisionMultiplier: BigNumberEthers.from(toPrecisionMultiplier)
-    //         });
-    //         return new BigNumber(out.toHexString());
-    //     }
-    //     const amountInWithFee = amountIn.multipliedBy(maxFee.minus(fee));
-    //     return amountInWithFee
-    //         .multipliedBy(reserveTo)
-    //         .dividedBy(new BigNumber(reserveFrom).multipliedBy(maxFee).plus(amountInWithFee));
-    // }
-    //
-    // private async getPoolPrecision(
-    //     address: string,
-    //     fromAddress: string
-    // ): Promise<{ fromPrecisionMultiplier: string; toPrecisionMultiplier: string }> {
-    //     const token0 = await this.web3Public.callContractMethod(
-    //         address,
-    //         syncSwapStablePool,
-    //         'token0',
-    //         []
-    //     );
-    //     const token0PM = await this.web3Public.callContractMethod(
-    //         address,
-    //         syncSwapStablePool,
-    //         'token0PrecisionMultiplier',
-    //         []
-    //     );
-    //     const token1PM = await this.web3Public.callContractMethod(
-    //         address,
-    //         syncSwapStablePool,
-    //         'token1PrecisionMultiplier',
-    //         []
-    //     );
-    //     return compareAddresses(token0, fromAddress)
-    //         ? {
-    //               fromPrecisionMultiplier: token0PM,
-    //               toPrecisionMultiplier: token1PM
-    //           }
-    //         : {
-    //               fromPrecisionMultiplier: token1PM,
-    //               toPrecisionMultiplier: token0PM
-    //           };
-    // }
-
     private async getAvailablePools(
         from: PriceTokenAmount<EvmBlockchainName>,
         toToken: PriceToken<EvmBlockchainName>
     ): Promise<RoutePools | null> {
-        return PoolFactory.fetchRoutePools(
+        return SyncSwapFactory.fetchRoutePools(
             from.address,
             toToken.address,
             this.walletAddress,
             '0x621425a1Ef6abE91058E9712575dcc4258F8d091',
             [
-                '0xf2dad89f2788a8cd54625c60b55cd3d2d0aca7cb',
-                '0x5b9f21d407f35b10cbfddca17d5d84b129356ea3'
+                '0xf2dad89f2788a8cd54625c60b55cd3d2d0aca7cb', // Stable
+                '0x5b9f21d407f35b10cbfddca17d5d84b129356ea3' // Regular
             ],
             [
-                '0x5aea5775959fbc2557cc8789bc1bf90a239d9a91',
-                '0x3355df6D4c9C3035724Fd0e3914dE96A5a83aaf4'
-            ].map(el => el.toLowerCase()),
+                '0x5aea5775959fbc2557cc8789bc1bf90a239d9a91', // WETH
+                '0x3355df6D4c9C3035724Fd0e3914dE96A5a83aaf4' // USDC
+            ].map(address => address.toLowerCase()),
             '0xbb05918e9b4ba9fe2c8384d223f0844867909ffb'
         );
     }
